requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
pynput>=1.7.6
pyperclip>=1.8.2
pyyaml>=6.0
soundfile>=0.12.0
